# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsInsight } from './analyticsInsight';
import type { AnalyticsKpi } from './analyticsKpi';
import type { AnalyticsReport } from './analyticsReport';
import type { ChannelRoas } from './channelRoas';
import type { FunnelStage } from './funnelStage';
import type { RevenuePoint } from './revenuePoint';
import type { TrafficSource } from './trafficSource';

export interface Analytics {
  kpis: AnalyticsKpi[];
  revenue: RevenuePoint[];
  channelRoas: ChannelRoas[];
  traffic: TrafficSource[];
  funnel: FunnelStage[];
  insight: AnalyticsInsight;
  reports: AnalyticsReport[];
}