# lib/api-zod/src/generated/types/analyticsKpi.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsKpiTrend } from './analyticsKpiTrend';

export interface AnalyticsKpi {
  label: string;
  value: string;
  change: string;
  trend: AnalyticsKpiTrend;
}