# lib/api-zod/src/generated/types/assistantConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantContext } from './assistantContext';
import type { AssistantMessage } from './assistantMessage';
import type { Guardrail } from './guardrail';

export interface AssistantConversation {
  messages: AssistantMessage[];
  context: AssistantContext[];
  guardrails: Guardrail[];
}