# lib/api-zod/src/generated/types/campaignSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignSummary {
  id: string;
  name: string;
  owner: string;
  ownerColor?: string;
  status: string;
  statusColor: string;
  progress: number;
  budgetTotal: number;
  budgetSpent: number;
  channels: string[];
}