# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttentionItem } from './attentionItem';
import type { CampaignSummary } from './campaignSummary';
import type { Insight } from './insight';
import type { KpiTile } from './kpiTile';
import type { Milestone } from './milestone';
import type { TaskRollup } from './taskRollup';

export interface DashboardSummary {
  kpis: KpiTile[];
  campaigns: CampaignSummary[];
  milestones: Milestone[];
  insights: Insight[];
  taskRollup: TaskRollup;
  attention: AttentionItem[];
}