# lib/api-zod/src/generated/types/guardrail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuardrailType } from './guardrailType';

export interface Guardrail {
  type: GuardrailType;
  title: string;
  description: string;
}